+++ gi_Player.h
@@ -9,11 +9,11 @@ struct gi_Player
 {
 	char m_name[GI_PLAYERNAME_MAX_SIZE];
 	GI_POSITION m_position;
-	int m_rawLevel;
-	int m_rawQST[GI_QST_NUM];
+	size_t m_rawLevel;
+	size_t m_rawQST[GI_QST_NUM];
 	float m_experience;
 	GI_UNIT m_unit;
-	int m_age;
+	size_t m_age;
 	gi_SpecialTeamsValues m_specialTeamsValues;
 	float m_level;
 	float m_QST[GI_QST_NUM];
